Changes to field lengths and names
Replaced any fields ending with underscore with a more suitable name. Email field was not remotely long enough to store most email addresses. The standard length of an email address is defined as 320 characters

Changed Credit Card Number to CHAR(16) instead of INT: 10^16 > 2^31, so card numbers would have caused overflow.
Split name field into first and last name which is standard.

Note that the specialization classes of User are made redundant by the UserType attribute. It would make more sense to have StAddress and CreditCard as Null-able fields, only filled when a user becomes a registered user. This was left untouched.