Added some new attributes :)
"Seats" is the number of seats available to be booked at the showtime. Through some wizardry, this will be synced to store the amount of seats available to be booked for a screening of a movie. "PremiumSeats" is the number of seats available to be booked by Registered Users only, in some predetermined timeframe (let's say 72 hrs) in advance. For the first 72 hours (let's say), PremiumSeats + Seats = TheatreRoom Capacity. This seating amount will be released after the first 72 hours, and the program will determine how many seats are available.